Experimental method to update Fiber payloads and a splitting example
Added method Fiber.updatePayloads that invokes a function to update
each payload in a fiber. This is useful for payload tranformations,
such as flattening or unflattening.

Added a Jupyter notebook that illustrates splitting a fiber using
both "positions" and "coordinates". Also illustrated splitting a
flattened fiber by position and then unflattening the split fiber.