Added methods that mutate a fiber tree
Created a set of methods that mutate the payloads at "self" or at some
depth below "self".

- splitUniformBelow()
- splitNonUniformBelow()
- splitEqualBelow()
- splitUnEqualBelow()
- swapRanksBelow()
- flattenBelow()
- unflattenBelow()

All of the above return None, since they modify the fiber tree at
"self".

To create the above:

Enhanced updatePayload() to take a "depth" keyword argument that
determines the depth in the tree at which the updates are to
occur.

Added a utility function that creates a closure around updatePayload()
with a update function from one of the several methods that return a
new fiber based on some operation on an existing fiber(s), such as
splitting, swapping or flattening.

Also added an IPython example of creating a 2D tiled tensor and a
start at a test suite.