Added {get,set}Color() method to tensors
Now TensorImage colors the scalar values in a Tensor with a
color assocaited with the Tensor (default "red"). New methods
{get,set}Color() were added. Note setColor() returns self, so
it can be used as in a chain.

So both coordinates and values weren't the same color changed
coordinate circles to be black. And made all labels white for
better contrast.

Updated Jupyter notebooks for spMspM and convolution to use
"standard" coloring scheme.

Added a .gitignore directive to ignore the tmp.mp4 file created
by TensorCanvas.