Added constructors for random fibers and tensors
Added constructors that given a set of rank_ids (tensor only), a
shape, a density for each level of the tree, an interval for each
value and an optional seed create a randomly populated tensor or
fiber.

Also added a __format__ function for tensors that just formats the
root fiber.