Reorganized handling of fiber default values
Moved handling of fiber default values to the owning rank from
fiber. This allows more consistent treatment of default value for both
intermediate and leaf level fibers in a tensor, i.e., the default is
either an empty fiber or 0 (or whatever was set using setDefault).

However, since raw fibers do not have an owning rank they will still
try to guess a default value. They also still support setDefault, but
its use is deprecated.

Added {set,get}Default as methods on Tensor. The set method controls
the default value returned as the payload of a leaf fiber.

Fiber.__or__ now properly returns (but doesn't insert) defaults for
the empty side of a co-iteration for intermediate levels of the tree.

Added associated tests.