Improve handling of tensor coordinates that are tuples
A variety of changes to better support tensor/fiber coordinates that
are tuples instead of just integers.

1) New Fiber flattening mode

Added a new flattening mode where the coordinates are of the flattened
ranks are "nested" into 2-element tuples instead of being flattened
into a single long tuple. This fixed swap when the coordinates in the
ranks being swapped were already tuples.

2) Tuple coordinates in UncompressedImage

Added support for showing coordinates of the top rank of a matrix that
were tuples in UncompressedImage.

TBD: Add support for coordinates in top rank of rank-3 tensors and
bottom rank of any tensor.

3) Simplify SpacetimeCanvas

Use fixed swapRanks method and native support for displaying tensor
coordinates that are tuples in UncompressedImage to greatly simplify
SpacetimeCanvas.