Added Fiber.getRankIds() and Rank.getRankIds() methods
Added methods for get the list of rank ids from a rank or from a
fiber. The fiber will try to interogate the rank for the
information (like it does for "shapes"), but will resort to a default
nameing scheme if the fiber is not part of a tensor.

These methods are useful for creating an empty new Tensor with the
same ranks as an existing tensor.