Added full-fledged CoordPayload class
Added a real class to implement the CoordPayload class, which is used
to represent an element of a fiber. This replaces the nametuple that
was used before, and currently supports iteration, addition,
multiplication and comparison operations.

Note: Iteration is used to automatically unpack the coordinate and
payload:

        coord, payload = CoordPayload(5, 10)

TBD: More operators and tests