Added methods Tensor.__getitem__() and Tensor.__setitem__()
Added some trampolines for __getitem__ and __setitem__ in Tensor,
which access the elements of the root fiber.

This allows one to write pretty natural code on uncompressed data and
take advantage of the display and animation tools that fiber trees
provide.

See examples/ipython/dense-dnn/dense-convolution.ipynb for an example.