Enhancements to the {Tensor,Fiber}.split*() methods
The Fiber.split*() methods now take a `depth` keyword argument to
specify the depth in the fibertree where the split should be
performed.

The {Tensor,Fiber}.split*() methods now take a `rankid` keyword
argument that specifies the name of the rank where the split should be
performed. The `rankid` keyword argument takes precedence over the
`depth` keyword argument.