Added iterators over a Fiber's shape
Implemented two methods to iterate over the **shape** of a Fiber,
Fiber.iterShape() and Fiber.iterShapeRef(), where the later will
create references to new **default** payloads for empty coordinates.

Note: I did not implement trampoline functions for these methods in
the Tensor class, since they should used so frequently as to warrent
them. We'll see if that is a good assumption.