Added {scalar,fiber}/fiber addition and multiplication operators
Added a series of operators on fibers that allow one to add or
multiply a scalar by each element of a fiber.

Note that because `Fiber.__add__()` and `Fiber.__iadd()` were already
defined as fiber concatentation, the new versions of those operators
look at the type of the `other` operand and do the concatentation if
it is a fiber and otherwise try to do the scalar/fiber addition.