Fix display of uncompressed tensors with tuple coordinates
This is a temporary workaround to enable the Jupyter notebook display
code to be able to display tensors with coordinates that are tuples.

In some past commits, the main fiber iterator code was motified in a
way that limits support to fibers with integer coordinates. This
commit adds some additional methods works for tuple coordinates and is
used by the uncompressed tensor display code. But the standard
iterators still do not work with tuple coordinates and need to be
updated.