Support for fully-populated tensors and fibers
Add support for fully-populated tensors and fibers means support for
the case where the `default` value for a tensor (or fiber) is
`None`. In that case, every coordinate of the object will have some
payload and thus `position` == `coordinate`. This is useful for
representing dense computations.

Changes include adding a method `Tensor.makePopulated()` to create a
tensor that is fully-populated with an initial value (normally zero)
and updates to the display functions to not give special treatment to
zeros when the default value is `None`.