Enhance Fiber.updateCoords() to allow changing rank_id/shape
To properly traverse every element of a fiber, i.e., in an
uncompressed fashion, the structure of the shape of the object must
match the structure of the coordinates of the object. For example if
the coordinates are integers or tuples then the shape must be an
integer or tuple, respectively. This is needed for uncompressed
displays to work.

The commit augments the `Fiber.updateCoords()` method to allow
changing the shape (and rank_id) when the update function makes a
change in the type of the coordinates, e.g., changing from integers to
tuples or vice versa.

That method also now checks that the type of the shape of the fiber
matches the type of the resulting coordinates.

Simiarly the workaround method Fiber.iterUncompressed() makes the same
check. Which has exposed the fact that we actually cannot properly
do uncompressed displays of flattened fibers.